Beginnings of a demo, also fixed a race condition with note playback
The way bhop was switching notes just happened to avoid this race
condition almost by chance, with a tiny window for problems. The
basic problem was writing a new zsaw_ptr without resetting the
zsaw_pos back to the beginning, which could sometimes result in
the next IRQ reading off the end of the array. If that caused it
to interpret a rate other than the valid $8 - $E range, then it
could push an OAM DMA too late and cause problems. Furthermore, by
reading corrupted sawtooth timing data, that note had a good chance
of being played incorrectly, or silencing itself. Fun times indeed.
Now fixed!