Add 4 more timbres AND improve performance, especially on high notes
The big useful observation was moving the rate setting logic earlier
in the IRQ, which greatly reduces the window of time before a timing
problem is introduced. This plays much more nicely with OAM DMA and
NMI interruptions, causing fewer artifacts and removing some of the
annoying buzzing that was present on high notes. It's still not
perfect, but it's *much* cleaner than before.