Include one more octave in the upper range, prioririze OAM DMA
The OAM DMA thing makes it just slightly less likely that we will
miss a timing window due to IRQ jitter.

The higher octave of notes is *not* especially well in tune, but
there are enough decently usable tones in that range that I can't
justify not including the set. It's a tiny splash of extra range
in certain keys.