Pull in updated library from the bhop project, mostly increasing safety
The most notable changes other than minor code cleanup are the timing
of sei/cli around the playback start routine, which dodges some issues
with incomplete state. Otherwise this is mostly for parity between
the two projects. Moving forward, z-saw will be authorative, it was
just easier to experiment and play with a real music engine driving
things.